New Login Page(s)
Added new login pages available under the NeoLogins folder, the main login page is still the original login page added in the original release. It is very simple to replace the login menus, just replace index.html with the one you want and rename the .txt to index.html. I plan on adding at least 1 more login menu and might replace the current main one with it, with the original still being archived possibly. These new login menus are not designed to be as modifiable as LetsChat itself as they were developed for LetsChat and are intended to be used along with LetsChat, hence these will not be released separately unlike some of the applications under the external services tab. Both the pages are one huge file that contain the HTML and CSS in one file but depend on LetsChat.js to function properly, both come with certain libraries pre-linked that are not required for the files to work but are there if needed while configuring.